feat: Implement ability to change shelf.
approach: the backend of BooksAPI is treated as the single source of
truth.

Updating a book is followed by updating our shelf state so that
Search and BookList maintain the same state of shelves at all times.

For this, the Search component needs to know which books are on the shelf, so that
while populating Book components, it can correctly set the
ShelfChanger's default selected value. On the other hand, the BookShelf
component already knows which book belongs to which shelf.